# THREESCALE-918: Optimize service contracts page loading