--- a/app/models/service_plan.rb
+++ b/app/models/service_plan.rb
@@ -10,6 +10,10 @@ class ServicePlan < Plan
 
   before_destroy :destroy_contracts
 
+  def self.provided_by(account)
+    Plan.by_type(self).issued_by(Service, account.service_ids)
+  end
+
   def provider_account
     service && service.account
   end