--- a/app/views/buyers/service_contracts/index.html.slim
+++ b/app/views/buyers/service_contracts/index.html.slim
@@ -37,7 +37,7 @@ table.data
       - unless @account
         th
           = sortable('accounts.org_name', 'Account')
-      - if current_account.multiservice?
+      - if @multiservice
         th
           label for="search_service_id"
             | Service
@@ -57,7 +57,7 @@ table.data
           th
             = s.text_field :account_query, size: 15
 
-          - if current_account.multiservice?
+          - if @multiservice
             th
               = s.collection_select :service_id, @services, :id, :name,
                                         { include_blank: true },
@@ -85,7 +85,7 @@ table.data
         - unless @account
           td
             = link_to_buyer_or_deleted contract.account
-        - if current_account.multiservice?
+        - if @multiservice
           td
             = link_to service.name, admin_service_path(service)
         td.plan
@@ -115,7 +115,7 @@ table.data
           td colspan="6"
             | No Service Subscriptions
       - else
-        = no_search_results(current_account.multiservice? ? 7 : 6)
+        = no_search_results(@multiservice ? 7 : 6)
 
 = will_paginate @service_contracts