--- a/test/unit/service_contract_test.rb
+++ b/test/unit/service_contract_test.rb
@@ -30,4 +30,28 @@ class ServiceContractTest < ActiveSupport::TestCase
     service = FactoryBot.create(:simple_service)
     assert ServiceContract.issued_by(service).count
   end
+
+  test '.provided_by scope' do
+    provider1 = FactoryBot.create(:simple_provider)
+    provider2 = FactoryBot.create(:simple_provider)
+    p1_service1 = FactoryBot.create(:simple_service, account: provider1)
+    p1_service2 = FactoryBot.create(:simple_service, account: provider1)
+    p2_service3 = FactoryBot.create(:simple_service, account: provider2)
+
+    # Default service plans are created on service creation, we destroy them for a clean comparison
+    p1_service1.service_plans.destroy_all
+    p1_service2.service_plans.destroy_all
+    p2_service3.service_plans.destroy_all
+
+    p1_s1_plan = FactoryBot.create(:simple_service_plan, issuer: p1_service1)
+    p1_s2_plan = FactoryBot.create(:simple_service_plan, issuer: p1_service2)
+    p2_s3_plan = FactoryBot.create(:simple_service_plan, issuer: p2_service3)
+
+    p1_contracts = FactoryBot.create_list(:simple_service_contract, 2, plan: p1_s1_plan) +
+                   FactoryBot.create_list(:simple_service_contract, 3, plan: p1_s2_plan)
+    p2_contracts = FactoryBot.create_list(:simple_service_contract, 4, plan: p2_s3_plan)
+
+    assert_same_elements p1_contracts, ServiceContract.provided_by(provider1).to_a
+    assert_same_elements p2_contracts, ServiceContract.provided_by(provider2).to_a
+  end
 end