**What this PR does / why we need it**:

The queries are not very efficient on the page, so it takes a long time to load the Service Subscriptions page - Audience > Accounts > Subscriptions, or `/buyers/service_contracts`.

**Which issue(s) this PR fixes** 

Service Subscriptions page takes too long to load: [THREESCALE-918](https://issues.redhat.com/browse/THREESCALE-918)

**Verification steps** 

Go to `/buyers/service_contracts` in the admin portal, and make sure that the loading times are reasonable (I used the staging database locally).

**Special notes for your reviewer**:

I used the SaaS Staging DB locally, because it is difficult to reproduce on an empty (local) DB, unless you fill the DB with a significant amount of data.

I used my test account, which has roughly 60 service subscriptions (3 pages), and the time of page loading are:

**Before**
```
Completed 200 OK in 28523ms (Views: 2404.0ms | ActiveRecord: 24932.4ms)
Completed 200 OK in 24114ms (Views: 2148.4ms | ActiveRecord: 20899.2ms)
Completed 200 OK in 28202ms (Views: 2305.9ms | ActiveRecord: 25045.8ms)
```

**After**
```
Completed 200 OK in 8785ms (Views: 1756.2ms | ActiveRecord: 6103.2ms)
Completed 200 OK in 3838ms (Views: 397.2ms | ActiveRecord: 3319.6ms)
Completed 200 OK in 8014ms (Views: 1284.7ms | ActiveRecord: 5808.6ms)
```

The query that has been improved significantly is the following, it was using nested queries, some of which were not relevant (querying for account plans and application plans for example, while they were not used).

**Before**
```
ServiceContract Load (15134.9ms)  SELECT  `cinstances`.* FROM `cinstances` WHERE `cinstances`.`type` IN ('ServiceContract') AND `cinstances`.`plan_id` IN (SELECT `plans`.`id` FROM `plans` WHERE (`plans`.`id` IN (SELECT `plans`.`id` FROM `plans` WHERE `plans`.`issuer_type` = 'Account' AND `plans`.`issuer_id` = 2445566778899 ORDER BY `plans`.`position` ASC) OR `plans`.`id` IN (SELECT `plans`.`id` FROM `plans` WHERE `plans`.`issuer_type` = 'Service' AND `plans`.`issuer_id` IN (2550000000001, 2550000000002, 2550000000003, 2550000000004, 2550000000005, 2550000000006, 2550000000007, 2550000000008, 2550000000009, 2550000000010, 2550000000011, 2550000000012, 2550000000013, 2550000000014, 2550000000015, 2550000000016, 2550000000017, 2550000000018, 2550000000019, 2550000000020, 2550000000021, 2550000000022, 2550000000023, 2550000000024, 2550000000025, 2550000000026, 2550000000027, 2550000000028, 2550000000029) ORDER BY `plans`.`position` ASC)) ORDER BY `plans`.`position` ASC) ORDER BY cinstances.created_at DESC LIMIT 20 OFFSET 40
```

**After**
```
ServiceContract Load (101.9ms)  SELECT  `cinstances`.* FROM `cinstances` WHERE `cinstances`.`type` IN ('ServiceContract') AND `cinstances`.`plan_id` IN (SELECT `plans`.`id` FROM `plans` WHERE `plans`.`type` IN ('ServicePlan') AND `plans`.`issuer_type` = 'Service' AND `plans`.`issuer_id` IN (2550000000001, 2550000000002, 2550000000003, 2550000000004, 2550000000005, 2550000000006, 2550000000007, 2550000000008, 2550000000009, 2550000000010, 2550000000011, 2550000000012, 2550000000013, 2550000000014, 2550000000015, 2550000000016, 2550000000017, 2550000000018, 2550000000019, 2550000000020, 2550000000021, 2550000000022, 2550000000023, 2550000000024, 2550000000025, 2550000000026, 2550000000027, 2550000000028, 2550000000029) ORDER BY `plans`.`position` ASC) ORDER BY cinstances.created_at DESC LIMIT 20 OFFSET 40
```